refactor(processors): unify import statements by consolidating pipeline imports into the main processor module
