+++ src/lerobot/datasets/pipeline_features.py
@@ -17,7 +17,7 @@ from typing import Any
 
 from lerobot.constants import ACTION, OBS_IMAGES, OBS_STATE
 from lerobot.datasets.utils import hw_to_dataset_features
-from lerobot.processor.pipeline import RobotProcessor
+from lerobot.processor import RobotProcessor
 
 
 def aggregate_pipeline_dataset_features(
