+++ src/lerobot/processor/batch_processor.py
@@ -17,10 +17,11 @@ from torch import Tensor
 
 from lerobot.configs.types import PolicyFeature
 from lerobot.constants import OBS_ENV_STATE, OBS_IMAGE, OBS_IMAGES, OBS_STATE
-from lerobot.processor.pipeline import (
+
+from .core import EnvTransition
+from .pipeline import (
     ActionProcessor,
     ComplementaryDataProcessor,
-    EnvTransition,
     ObservationProcessor,
     ProcessorStep,
     ProcessorStepRegistry,
