+++ src/lerobot/utils/train_utils.py
@@ -32,7 +32,7 @@ from lerobot.datasets.utils import load_json, write_json
 from lerobot.optim.optimizers import load_optimizer_state, save_optimizer_state
 from lerobot.optim.schedulers import load_scheduler_state, save_scheduler_state
 from lerobot.policies.pretrained import PreTrainedPolicy
-from lerobot.processor.pipeline import RobotProcessor
+from lerobot.processor import RobotProcessor
 from lerobot.utils.random_utils import load_rng_state, save_rng_state
