+++ tests/processor/test_batch_processor.py
@@ -22,9 +22,7 @@ import pytest
 import torch
 
 from lerobot.constants import OBS_ENV_STATE, OBS_IMAGE, OBS_IMAGES, OBS_STATE
-from lerobot.processor import ProcessorStepRegistry, RobotProcessor
-from lerobot.processor.batch_processor import ToBatchProcessor
-from lerobot.processor.pipeline import TransitionKey
+from lerobot.processor import ProcessorStepRegistry, RobotProcessor, ToBatchProcessor, TransitionKey
 
 
 def create_transition(
