+++ tests/processor/test_pi0_processor.py
@@ -29,9 +29,9 @@ from lerobot.processor import (
     NormalizerProcessor,
     RenameProcessor,
     ToBatchProcessor,
+    TransitionKey,
     UnnormalizerProcessor,
 )
-from lerobot.processor.pipeline import TransitionKey
 
 
 def create_transition(observation=None, action=None, **kwargs):
